+++ src/columns/file_name.rs
@@ -1,0 +1,39 @@
+use crate::process::ProcessInfo;
+use crate::{column_default, Column};
+use std::cmp;
+use std::collections::HashMap;
+
+pub struct FileName {
+    header: String,
+    unit: String,
+    fmt_contents: HashMap<i32, String>,
+    raw_contents: HashMap<i32, String>,
+    width: usize,
+}
+
+impl FileName {
+    pub fn new(header: Option<String>) -> Self {
+        let header = header.unwrap_or_else(|| String::from("FileName"));
+        let unit = String::new();
+        FileName {
+            fmt_contents: HashMap::new(),
+            raw_contents: HashMap::new(),
+            width: 0,
+            header,
+            unit,
+        }
+    }
+}
+
+#[cfg(any(target_os = "linux", target_os = "android"))]
+impl Column for FileName {
+    fn add(&mut self, proc: &ProcessInfo) {
+        let raw_content = proc.curr_proc.stat().comm.clone();
+        let fmt_content = raw_content.clone();
+
+        self.fmt_contents.insert(proc.pid, fmt_content);
+        self.raw_contents.insert(proc.pid, raw_content);
+    }
+
+    column_default!(String);
+}